% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getFunctions.R
\docType{methods}
\name{getCase}
\alias{getCase}
\alias{getCase,MultiSynth-method}
\title{Extract prep matrices and fit on one of the cases in a MultiSynth analysis}
\usage{
getCase(input, case)

\S4method{getCase}{MultiSynth}(input, case)
}
\arguments{
\item{input}{An object of class "MultiSynth"}

\item{case}{The name of one of the cases.  For a placebo analysis, this is the unit name of interest.  For leave-one-out analyses, this is "minus <<excluded unit/covaraite>>".  To see the full list of names in a MultiSynth analaysis, one can call \code{names(fitMultiSynth.out@preps)}.}
}
\value{
A list containing the datapre matrices and synth output for this case
}
\description{
This function allows the user to extract the objects necessary to investigate any single case in a MultiSynth anlaysis.  It returns both the dataprep matrices and the output of the call to \code{\link{synth}}.  Together these can be used by various other functions, such as \code{\link{synth.tab}}, \code{\link{path.plot}}, \code{\link{SynthErrorRatios}}, \code{\link{SynthMeanEffect}}, and \code{\link{gaps.plot}}.
}
\author{
Dalston G. Ward: \email{ward.dalston@wustl.edu}
}
\seealso{
\code{\link{MultiSynth}}

\code{\link{getInput}}

\code{\link{getPreps}}

\code{\link{getFits}}

\code{\link{getTreated}}

\code{\link{getTreatmentTime}}

\code{\link{getStats}}
}

