% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getFunctions.R
\docType{methods}
\name{getFits}
\alias{getFits}
\alias{getFits,MultiSynth-method}
\title{Extract the synthetic control fits from a MultiSynth object}
\usage{
getFits(input)

\S4method{getFits}{MultiSynth}(input)
}
\arguments{
\item{input}{An object of class "MultiSynth"}
}
\value{
A list containing the output from \code{\link{synth}} for each of the placebo/leave-one-out cases
}
\description{
Returns a list containing the output from \code{\link{synth}} for each of the placebo/leave-one-out cases.
}
\author{
Dalston G. Ward: \email{ward.dalston@wustl.edu}
}
\seealso{
\code{\link{MultiSynth}}

\code{\link{getInput}}

\code{\link{getPreps}}

\code{\link{getTreated}}

\code{\link{getTreatmentTime}}

\code{\link{getCase}}

\code{\link{getStats}}
}

