% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getFunctions.R
\name{getPreps}
\alias{getPreps}
\title{Extract the data matrices for the placebo/leave-one-out cases}
\usage{
getPreps(input)
}
\arguments{
\item{input}{An object of class "MultiSynth"}
}
\value{
A list of lists containing the six matrices for each placebo/leave-one-out case.
}
\description{
Extracts from a MultiSynth object the named list containing the X0, X1, Z0, Z1, Y0plot, and Y1plot matrices for each of the placebo/leave-one-out cases. Each element of this list is a list containing the six relevent matrices for each placebo/leave-one-out case.
}
\author{
Dalston G. Ward: \email{ward.dalston@wustl.edu}
}
\seealso{
\code{\link{MultiSynth}}

\code{\link{getInput}}

\code{\link{getFits}}

\code{\link{getTreated}}

\code{\link{getTreatmentTime}}

\code{\link{getCase}}

\code{\link{getStats}}
}

