% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getFunctions.R
\docType{methods}
\name{getStats}
\alias{getStats}
\alias{getStats,MultiSynth-method}
\title{Extract the Statistics for Estimated Synthetic Control Fits in a MultiSynth Analysis}
\usage{
getStats(input)

\S4method{getStats}{MultiSynth}(input)
}
\arguments{
\item{input}{An object of class "MultiSynth"}
}
\value{
A matrix containing the estimated statistics for each case
}
\description{
Returns a matrix containing the pre-treatment RMSPE, post-treatment RMSPE, post- to pre-treatment RMSPE ratio, covariate loss, and ATE for each case in a MultiSynth analysis.
}
\author{
Dalston G. Ward: \email{ward.dalston@wustl.edu}
}
\seealso{
\code{\link{MultiSynth}}

\code{\link{getInput}}

\code{\link{getPreps}}

\code{\link{getFits}}

\code{\link{getTreated}}

\code{\link{getTreatmentTime}}

\code{\link{getCase}}
}

