% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getFunctions.R
\docType{methods}
\name{getTreated}
\alias{getTreated}
\alias{getTreated,MultiSynth-method}
\title{Extract Name and Number of Treated Unit in MultiSynth Analysis}
\usage{
getTreated(input)

\S4method{getTreated}{MultiSynth}(input)
}
\arguments{
\item{input}{An object of class "MultiSynth"}
}
\value{
A character vector containing the name and number of the treated unit
}
\description{
Returns a vector containing the name and unit number of the treated unit in a MultiSynth object.
}
\details{
If the user does not supply an argument to \code{unit.names.variable} in the original call to \code{\link{dataprep}}, then this function returns the unit number in both elements of the output vector.
}
\author{
Dalston G. Ward: \email{ward.dalston@wustl.edu}
}
\seealso{
\code{\link{MultiSynth}}

\code{\link{getInput}}

\code{\link{getPreps}}

\code{\link{getFits}}

\code{\link{getTreatmentTime}}

\code{\link{getCase}}

\code{\link{getStats}}
}

