% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getFunctions.R
\name{getTreatmentTime}
\alias{getTreatmentTime}
\title{Extract information about when treatment occurs MultiSynth analysis}
\usage{
getTreatmentTime(input)
}
\arguments{
\item{input}{An object of class "MultiSynth"}
}
\value{
A numeric vector of length one containing the time treatment is administered
}
\description{
Returns the user-given treatement time point in a MultiSynth object
}
\author{
Dalston G. Ward: \email{ward.dalston@wustl.edu}
}
\seealso{
\code{\link{MultiSynth}}

\code{\link{getInput}}

\code{\link{getPreps}}

\code{\link{getFits}}

\code{\link{getTreated}}

\code{\link{getCase}}

\code{\link{getStats}}
}

