% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MultiSynthPrepFunctions.R
\name{LOOcovariatesMSPrep}
\alias{LOOcovariatesMSPrep}
\title{Prepare data matrices for a Leave-One-Out Covariates MultiSynth Analysis}
\usage{
LOOunitsMSPrep(input)
}
\arguments{
\item{input}{The output from a call to \code{\link{dataprep}}}
}
\value{
A list of lists, with each list containing the elements
  \itemize{
  \item{X0}{A matrix of predictor values for donor pool units}
  \item{X1}{A vector of predictor values for the treated unit}
  \item{Z0}{A matrix of pre-treatment outcome vaules for donor pool units}
  \item{Z1}{A vector of pre-treatment outcome values for the treated unit}
  \item{Y0plot}{A matrix of outcome values for the donor pool units}
  \item{Y1plot}{A vector of outcome values for the treated unit}
  }
}
\description{
This function creates the necessary sets of data matrices for running a leave-one-out covariates MultiSynth analysis.  For each covariate used in the main analysis, it creates an X0 and X1 matrix excluding a given covariate, copies the original Z0, Z1, Y0plot, Y1plot matrices, and creates a list containing these six matrices. It also returns the original set of matrices, with the full covariate set. This function takes as input the list returned by \code{\link{dataprep}}.  This function is not intended to be called directly by the user, but rather is called indirectly by \code{\link{fitMultiSynth}}.
}
\details{
The output of this function is a named list, with each element being a list of the 6 matrices necessary for running \code{\link{synth}} and plotting its output. The names of this list are "Full Covariate set" for the full set, and "Minus <<covariate name>>" for the leave-one-out cases.
}
\author{
Dalston G. Ward: \email{ward.dalston@wustl.edu}
}
\references{
\itemize{
\item Abadie, A., Diamond, A., Hainmueller, J. (2010). Synthetic Control Methods for Comparative Case Studies: Estimating the Effect of California's Tobacco Control Program. \emph{Journal of the American Statistical Association} 105 (490) 493-505.
\item Abadie, A., Diamond, A., Hainmueller, J. (2015). Comparative Politics and the Synthetic Control Method.  \emph{American Journal of Political Science} 59 (2) 495-510
}
}
\seealso{
\code{\link{fitMultiSynth}}

\code{\link{PlaceboMSPrep}}

\code{\link{LOOunitsMSPrep}}
}

