% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MultiSynthPrepFunctions.R
\name{LOOunitsMSPrep}
\alias{LOOunitsMSPrep}
\title{Prepare Data for a Leave-One-Out Units MultiSynth Analysis}
\usage{
LOOunitsMSPrep(input)
}
\arguments{
\item{input}{The output from a call to \code{\link{dataprep}}}
}
\value{
A list of lists, with each list containing the elements
  \itemize{
  \item{X0}{A matrix of predictor values for donor pool units}
  \item{X1}{A vector of predictor values for the treated unit}
  \item{Z0}{A matrix of pre-treatment outcome vaules for donor pool units}
  \item{Z1}{A vector of pre-treatment outcome values for the treated unit}
  \item{Y0plot}{A matrix of outcome values for the donor pool units}
  \item{Y1plot}{A vector of outcome values for the treated unit}
  }
}
\description{
This function creates the necessary sets of matrices for running a leave-one-out units MultiSynth analysis.  For each unit in the donor pool, it creates the X0, X1, Z0, Z1, Y0plot, and Y1plot matrices with the given unit removed from the donor pool.  It also returns these six matrices for original case, with the full donor pool.  This function takes as input the output of \code{\link{dataprep}}.  This function is not intended to be called directly by the user, but rather is called indirectly by \code{\link{fitMultiSynth}}.
}
\details{
The output of this function is a named list, with each element being a list of the 6 matrices necessary for running \code{\link{synth}} and plotting its output.  If the user supplies character unit names in the call to dataprep used as the input, then the names of the output list will be these names.  Otherwise, unit numbers are the output list names.
}
\author{
Dalston G. Ward: \email{ward.dalston@wustl.edu}
}
\references{
\itemize{
\item Abadie, A., Diamond, A., Hainmueller, J. (2010). Synthetic Control Methods for Comparative Case Studies: Estimating the Effect of California's Tobacco Control Program. \emph{Journal of the American Statistical Association} 105 (490) 493-505.
\item Abadie, A., Diamond, A., Hainmueller, J. (2015). Comparative Politics and the Synthetic Control Method.  \emph{American Journal of Political Science} 59 (2) 495-510
}
}
\seealso{
\code{\link{fitMultiSynth}}

\code{\link{PlaceboMSPrep}}

\code{\link{LOOcovariatesMSPrep}}
}

