% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MultiSynthPlotPrepFunctions.R
\name{MultiSynthPathPlotPrep}
\alias{MultiSynthPathPlotPrep}
\title{Extract values necessary for plotting the outcome trajectory for placebo/leave-one-out Synthetic fits}
\usage{
MultiSynthPathPlotPrep(MultiSynthPrep_obj, MultiSynth_obj)
}
\arguments{
\item{MultiSynthPrep_obj}{A list the dataprep matrices for each placebo/leave-one-out analysis.}

\item{MultiSynth_obj}{A list containing the output from \code{synth} for each placebo/leave-one-out analysis.}
}
\description{
This function extracts the values necessary for plotting the outcome trajectory for placebo/leave-one-out synthetic controls from a list of dataprep matrices and a list of outputs from \code{\link{synth}}, such as are found in \code{\link{MultiSynth}} objects.  It is not intended to be called directly by the user, but is rather called as part of the the \code{\link{MultiSynth}} methods for \code{\link{path.plot}}.
}
\details{
The function is called as part of \code{\link{path.plot,MultiSynth-method}}.  It is designed to take the the lists in the preps and fits slot of a \code{\link{MultiSynth}} object as its main input. If one wants to use a custom input, these should be lists containing dataprep objects and the output of synth.
}
\author{
Dalston G. Ward: \email{ward.dalston@wustl.edu}
}
\references{
\itemize{
\item Abadie, A., Diamond, A., Hainmueller, J. (2010). Synthetic Control Methods for Comparative Case Studies: Estimating the Effect of California's Tobacco Control Program. \emph{Journal of the American Statistical Association} 105 (490) 493-505.
\item Abadie, A., Diamond, A., Hainmueller, J. (2015). Comparative Politics and the Synthetic Control Method.  \emph{American Journal of Political Science} 59 (2) 495-510
}
}
\seealso{
\code{\link{MultiSynth}}

\code{\link{path.plot,MultiSynth-method}}

\code{\link{MultiSynthPlotLines}}

\code{\link{MultiSynthGapPlotPrep}}
}

