% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MultiSynthPlotPrepFunctions.R
\name{MultiSynthPlotLines}
\alias{MultiSynthPlotLines}
\title{Add Lines Representing Placebo/Leave-One-Out Anlayses to a Path or Gaps plot}
\usage{
MultiSynthPlotLines(MultiSynthPlotPrep_obj, col = "gray48", lwd = 3/4,
  cex = 1, t = "l")
}
\arguments{
\item{MultiSynthPlotPrep_obj}{A list containing the x and y values fore each of the lines to be plotted}

\item{col}{Color to plot lines.  Defaults to gray.}

\item{lwd}{Width of plotted lines. Defaults to 3/4}

\item{cex}{Expansion parameter for plotting. Defaults to 1.}

\item{t}{Plotting type.  Defaults to "l"}
}
\description{
This function plots the lines for the placebo/leave-one-out synthetic controls when \code{\link{path.plot}} or \code{\link{gaps.plot}} is called on a \code{\link{MultiSynth}} object.  It is not intended to be called directly by the user.
}
\details{
The function is called as part of \code{\link{path.plot,MultiSynth-method}} and \code{\link{gaps.plot,MultiSynth-method}}.  It is designed to take the output of \code{\link{MultiSynthPathPlotPrep}} or \code{\link{MultiSynthGapPlotPrep}} as its main input. If one wants to use a custom input object, the input should be a list with each element being a matrix with two columns. One of the columns, contianing x values, should be called "x.info", the other, with y values, called "y.info".
}
\author{
Dalston G. Ward: \email{ward.dalston@wustl.edu}
}
\references{
\itemize{
\item Abadie, A., Diamond, A., Hainmueller, J. (2010). Synthetic Control Methods for Comparative Case Studies: Estimating the Effect of California's Tobacco Control Program. \emph{Journal of the American Statistical Association} 105 (490) 493-505.
\item Abadie, A., Diamond, A., Hainmueller, J. (2015). Comparative Politics and the Synthetic Control Method.  \emph{American Journal of Political Science} 59 (2) 495-510
}
}
\seealso{
\code{\link{MultiSynth}}

\code{\link{path.plot,MultiSynth-method}}

\code{\link{gaps.plot,MultiSynth-method}}

\code{\link{MultiSynthGapPlotPrep}}

\code{\link{MultiSynthPathPlotPrep}}
}

