% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MultiSynthPathPlot.R
\docType{methods}
\name{path.plot,MultiSynth-method}
\alias{path.plot,MultiSynth-Method}
\alias{path.plot,MultiSynth-method}
\title{Path plots for MultiSynth Objects}
\usage{
\S4method{path.plot}{MultiSynth}(synth.res = NA, dataprep.res = NA,
  tr.intake = NA, Ylab = c("Y Axis"), Xlab = c("Time"), Ylim = NA,
  Legend = c("Treated", "Synthetic", "Alternative Synthetics"),
  Legend.position = c("topleft"), Main = NA, Z.plot = FALSE)
}
\arguments{
\item{synth.res}{An object of class \code{\link{MultiSynth}}}

\item{dataprep.res}{NA.  Not used for MultiSynth path plotting, should be left as NA.}

\item{tr.intake}{NA. Not used for MultiSynth path plotting, should be left as NA.}

\item{Ylab}{Optional label for Y axis}

\item{Xlab}{Optional label for X axis}

\item{Ylim}{Optional Y axis limits}

\item{Legend}{Optional legend text.  If given as NA no legend is plotted.}

\item{Legend.position}{Optional legend position.  See \code{\link{legend}} for details.}

\item{Main}{Optional main title}

\item{Z.plot}{Logical.  If TRUE only the pre-treatment period is plotted.}
}
\description{
Plots the outcome trajectory of the outcome variable for the treated unit, the main synthetic control, and the place/leave-one-out synthetic controls created using \code{\link{fitMultiSynth}}.  The treated unit's trajectory is plotted as a solid line, the main synthetic control as a dashed line, and the MultiSynth controls as light gray solid lines.
}
\details{
The user should give an object of class \code{\link{MultiSynth}} as the argument to synth.res, and leave dataprep.res and tr.intake as NA.

This method for path.plot is generally designed to work with leave-one-out analayses.  It will work for placebo anlayses, but the interpretation of the plot is less clear.
}
\examples{
\dontrun{
##Example: Hainmueller and Diamond's Toy panel dataset

 #load data
data(synth.data)

## create matrices from panel data that provide inputs for fitMultiSynth()
dataprep.out<-
 dataprep(
   foo = synth.data,
   predictors = c("X1", "X2", "X3"),
   predictors.op = "mean",
   dependent = "Y",
   unit.variable = "unit.num",
   time.variable = "year",
   special.predictors = list(
     list("Y", 1991, "mean"),
     list("Y", 1985, "mean"),
     list("Y", 1980, "mean")
   ),
   treatment.identifier = 7,
   controls.identifier = c(29, 2, 13, 17, 32, 38),
   time.predictors.prior = c(1984:1989),
   time.optimize.ssr = c(1984:1990),
   unit.names.variable = "name",
   time.plot = 1984:1996
 )

 ##Fit a MultiSynth
 fitMultiSynth.out <- fitMultiSynth(dataprep.out, type = "units", treatment_time = 1991)

path.plot(fitMultiSynth.out)

}
}
\author{
Dalston G. Ward: \email{ward.dalston@wustl.edu}
}
\references{
\itemize{
\item Abadie, A., Diamond, A., Hainmueller, J. (2010). Synthetic Control Methods for Comparative Case Studies: Estimating the Effect of California's Tobacco Control Program. \emph{Journal of the American Statistical Association} 105 (490) 493-505.
\item Abadie, A., Diamond, A., Hainmueller, J. (2015). Comparative Politics and the Synthetic Control Method.  \emph{American Journal of Political Science} 59 (2) 495-510
}
}
\seealso{
\code{\link{fitMultiSynth}}

\code{\link{MultiSynth}}

\code{\link{plot,MultiSynth-method}}

\code{\link{summary,MultiSynth-method}}

\code{\link{gaps.plot,MultiSynth-method}}

\code{\link{path.plot}}
}

