% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotMultiSynth.R
\docType{methods}
\name{plot,LOOunitsMS,ANY-method}
\alias{plot,LOOcovariatesMS,ANY-method}
\alias{plot,LOOunitsMS,ANY-method}
\alias{plot,MultiSynth-method}
\alias{plot,PlaceboMS,ANY-method}
\title{Plot MultiSynth objects}
\usage{
\S4method{plot}{LOOunitsMS,ANY}(x, y, quantity = "Ratios", main = NA, ...)

\S4method{plot}{LOOcovariatesMS,ANY}(x, y, quantity = "Ratios", main = NA,
  ...)

\S4method{plot}{PlaceboMS,ANY}(x, y, quantity = "Ratios", main = NA, ...)
}
\arguments{
\item{x}{A MultiSynth object}

\item{y}{Should be left blank; not used in this method}

\item{quantity}{The quantity to plot.  Can be "Ratios" for RMSPE ratio, "Pre" for pretreatment RMSPE, "Post" for posttreatment RMSPE, "ATE" for ATEs, or "Cov" for covariate balance.  Defaults to "Ratios".}

\item{main}{Optional main title. Defaults to "<<Quantity to plot>> for <<type of MultiSynth>> Analysis of <<treated unit name>>".}

\item{...}{further arguments passed on to the default method of plot}
}
\description{
This function plots the distribution of values for the statistics calculated in a MultiSynth analysis.  It can plot five quantities: pretreatment RMSPEs, posttreatment RMSPEs, post-to-pre-treatment RMSPE ratios, ATEs, and covariate loss.  The user chooses which quantity to plot through the arguement quantity.




}
\details{
The function first sorts the quantity of interest, and then plots on the x-axis the values of the quantity for each case in the MultiSynth analysis.  Axis labels correspond to the names of the placebo/leave-one-out cases.  When unit or covariate names are very long, they may not fit in the plotting space.  This can be corrected by adjusting the plotting margins via \code{\link{par}}. The actual treated unit or the analysis with the full donor pool/covariate set is plotted in red.

The three different subclasses of MultiSynth objects (PlaceboMS, LOOunitsMS, and LOOcovariatesMS) each have their own method for plot.  The only difference between these three methods, however, is the text printed by default for the main plot title.
}
\examples{
\dontrun{
##Example: Hainmueller and Diamond's Toy panel dataset

#load data
data(synth.data)

## create matrices from panel data that provide inputs for fitMultiSynth()
dataprep.out<-
 dataprep(
   foo = synth.data,
   predictors = c("X1", "X2", "X3"),
   predictors.op = "mean",
   dependent = "Y",
   unit.variable = "unit.num",
   time.variable = "year",
   special.predictors = list(
     list("Y", 1991, "mean"),
     list("Y", 1985, "mean"),
     list("Y", 1980, "mean")
   ),
   treatment.identifier = 7,
   controls.identifier = c(29, 2, 13, 17, 32, 38),
   time.predictors.prior = c(1984:1989),
   time.optimize.ssr = c(1984:1990),
   unit.names.variable = "name",
   time.plot = 1984:1996
 )

 fitMultiSynth.out <- fitMultiSynth(dataprep.out, treatment_time = 1991)

 plot(fitMultiSynth.out)

 plot(fitMultiSynth.out, quantity = "ATE")
 }
}
\author{
Dalston G. Ward: \email{ward.dalston@wustl.edu}
}
\references{
\itemize{
\item Abadie, A., Diamond, A., Hainmueller, J. (2010). Synthetic Control Methods for Comparative Case Studies: Estimating the Effect of California's Tobacco Control Program. \emph{Journal of the American Statistical Association} 105 (490) 493-505.
\item Abadie, A., Diamond, A., Hainmueller, J. (2015). Comparative Politics and the Synthetic Control Method.  \emph{American Journal of Political Science} 59 (2) 495-510
}
}
\seealso{
\code{\link{fitMultiSynth}}

\code{\link{MultiSynth}}
}

