% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PPA}
\alias{PPA}
\title{Characteristics of 1000 People in Pennsylvania.}
\format{A data frame with 1000 rows and 7 variables:
\describe{
  \item{age}{age of respondent, in years, ("AGEP")}
  \item{sex}{sex of respondent, ("SEX")}
  \item{race}{recoded detailed race code, ("RAC1P")}
  \item{marriage}{married/spouse present/spouse absent, ("MSP")}
  \item{emp}{employment status recode, ("ESR")}
  \item{vet}{veteran period of service, ("VPS")}
  \item{age17plus}{age >= 17 indicator}
}}
\source{
\url{https://factfinder.census.gov/faces/tableservices/jsf/pages/productview.xhtml?pid=ACS_pums_csv_2017&prodType=document}
}
\usage{
PPA
}
\description{
A dataset containing some variables about 1000 people in Pennsylvania.  This is a subset of the 2017 ACS PUMS data with one indicator variable added.
}
\keyword{datasets}
