% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PPAm5}
\alias{PPAm5}
\title{A list containing 5 partially synthetic data sets.}
\format{5 data frames with 1000 rows and 7 variables:
\describe{
  \item{age}{age of respondent, in years, ("AGEP")}
  \item{sex}{sex of respondent, ("SEX")}
  \item{race}{recoded detailed race code, ("RAC1P")}
  \item{marriage}{married/spouse present/spouse absent, ("MSP")}
  \item{emp}{employment status recode, ("ESR")}
  \item{vet}{veteran period of service, ("VPS")}
  \item{age17plus}{age >= 17 indicator}
}}
\source{
\url{https://factfinder.census.gov/faces/tableservices/jsf/pages/productview.xhtml?pid=ACS_pums_csv_2017&prodType=document}
}
\usage{
PPAm5
}
\description{
This is a list that has the 5 partially synthetic versions of PPA (PPAps1 - PPAps5).
}
\keyword{datasets}
