% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PPAps2}
\alias{PPAps2}
\title{Characteristics of 1000 People in Pennsylvania, partially synthetic (set 2).}
\format{A data frame with 1000 rows and 7 variables:
\describe{
  \item{age}{age of respondent, in years, ("AGEP")}
  \item{sex}{sex of respondent, ("SEX")}
  \item{race}{recoded detailed race code, ("RAC1P")}
  \item{marriage}{married/spouse present/spouse absent, ("MSP")}
  \item{emp}{employment status recode, ("ESR")}
  \item{vet}{veteran period of service, ("VPS")}
  \item{age17plus}{age >= 17 indicator}
}}
\source{
\url{https://factfinder.census.gov/faces/tableservices/jsf/pages/productview.xhtml?pid=ACS_pums_csv_2017&prodType=document}
}
\usage{
PPAps2
}
\description{
This is a version of the PPA data set that is partially synthetic.  Some of the values of "sex", "marriage", and "age17plus" were imputed.
}
\keyword{datasets}
