% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/printSynthErrorRatio.R
\name{print.SynthErrorRatios}
\alias{print.SynthErrorRatios}
\title{Print the RMSPE values for a Synthetic Control Analysis}
\usage{
\method{print}{SynthErrorRatios}(x, digits = 3, ...)
}
\arguments{
\item{x}{An object of class SynthErrorRatios}

\item{digits}{The number of digits to print, defaults to 3}

\item{...}{Additional arguments passed to other functions}
}
\description{
This function prints the pretreatment root mean square predictive error RMSPE values, posttreatment RMSPE, and the ratio of these two values in a synthetic control analysis, labelling each statistic.
}
\details{
This function is called silently anytime the user sends an object of class SynthErrorRatios directly to the console.
}
\author{
Dalston G. Ward: \email{ward.dalston@wustl.edu}
}
\references{
\itemize{
\item Abadie, A., Diamond, A., Hainmueller, J. (2010). Synthetic Control Methods for Comparative Case Studies: Estimating the Effect of California's Tobacco Control Program. \emph{Journal of the American Statistical Association} 105 (490) 493-505.
\item Abadie, A., Diamond, A., Hainmueller, J. (2015). Comparative Politics and the Synthetic Control Method.  \emph{American Journal of Political Science} 59 (2) 495-510
}
}
\seealso{
\code{\link{SynthRMSPE}}

\code{\link{synth}}

\code{\link{dataprep}}
}

