% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/showMultiSynth.R
\docType{methods}
\name{show,PlaceboMS-method}
\alias{show,LOOcovariatesMS-method}
\alias{show,LOOunitsMS-method}
\alias{show,MultiSynth-method}
\alias{show,PlaceboMS-method}
\title{Show objects of class "MultiSynth"}
\usage{
\S4method{show}{PlaceboMS}(object)

\S4method{show}{LOOunitsMS}(object)

\S4method{show}{LOOcovariatesMS}(object)
}
\arguments{
\item{object}{An object of class MultiSynth.  Can be a placebo analysis or either type of leave-one-out analysis.}
}
\description{
Display the basic information about an object of class "MultiSynth".  This prints to the console the type of MultiSynth analysis, information about the treated unit, the donor pool, and covariate set, and the time periods used in fitting the analysis.  The function is invoked by automatic printing (when an object of class "MultiSynth" is sent to the console, or not assigned after a call to \code{\link{fitMultiSynth}}).
}
\details{
There are methods of show defined for each of the three subclasses of MultiSynth.  The only differences between the show methods for these classes is the title of the output printed to the console.

If the user does not provide character unit names in the \code{\link{dataprep}} output used in the call to \code{\link{fitMultiSynth}} then unit numbers are used in the place of unit names for this method.
}
\author{
Dalston G. Ward: \email{ward.dalston@wustl.edu}
}
\references{
\itemize{
\item Abadie, A., Diamond, A., Hainmueller, J. (2010). Synthetic Control Methods for Comparative Case Studies: Estimating the Effect of California's Tobacco Control Program. \emph{Journal of the American Statistical Association} 105 (490) 493-505.
\item Abadie, A., Diamond, A., Hainmueller, J. (2015). Comparative Politics and the Synthetic Control Method.  \emph{American Journal of Political Science} 59 (2) 495-510
}
}
\seealso{
\code{\link{fitMultiSynth}}

\code{\link{MultiSynth}}
}

