% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summaryMultiSynth.R
\docType{methods}
\name{summary,PlaceboMS-method}
\alias{summary,LOOcovariatesMS-method}
\alias{summary,LOOunitsMS-method}
\alias{summary,MultiSynth-method}
\alias{summary,PlaceboMS-method}
\title{Summary of a MultiSynth Analysis}
\usage{
\S4method{summary}{PlaceboMS}(object, digits = 3)

\S4method{summary}{LOOunitsMS}(object, digits = 3)

\S4method{summary}{LOOcovariatesMS}(object, digits = 3)
}
\arguments{
\item{object}{An object of class MultiSynth}

\item{digits}{The number of digits to print.  Defaults to 3.}
}
\value{
A matrix containing all of the statistics printed to the console.
}
\description{
This method provides a summary of the results of a MultiSynth analysis.  It prints to the console the values of pre- and post-treatment RMSPEs, their ratios, ATEs, and covariate loss statistics for each case fitted in a MultiSynth analysis.  For placebo analyses, it also prints the p-value of obtaining an RMSPE ratio as large as the treated unit when treatment is randomly assigned to any unit in the donor pool. A matrix containing all of these statistics is silently returned.
}
\details{
Together, the quantities reported by this method allow the user to determine the quality of fit, robustness, and significance of a synthetic control analysis.  For details, see Abadie, Diamond, Hainmueller (2010, 2015)

Methods for summary are defined for each of the three MultiSynth subclasses.  However, there are only very minor differences between the three methods.  First, the title of the output printed varies depending on the type of MultiSynth analysis.  Second, for placebo analyses, the p-value for the RMSPE ratio is printed to the console.  Otherwise, the three methods are exactly the same.
}
\examples{
\dontrun{
##Example: Hainmueller and Diamond's Toy panel dataset

#load data
data(synth.data)

## create matrices from panel data that provide inputs for fitMultiSynth()
dataprep.out<-
 dataprep(
   foo = synth.data,
   predictors = c("X1", "X2", "X3"),
   predictors.op = "mean",
   dependent = "Y",
   unit.variable = "unit.num",
   time.variable = "year",
   special.predictors = list(
     list("Y", 1991, "mean"),
     list("Y", 1985, "mean"),
     list("Y", 1980, "mean")
   ),
   treatment.identifier = 7,
   controls.identifier = c(29, 2, 13, 17, 32, 38),
   time.predictors.prior = c(1984:1989),
   time.optimize.ssr = c(1984:1990),
   unit.names.variable = "name",
   time.plot = 1984:1996
 )

 fitMultiSynth.out <- fitMultiSynth(dataprep.out, treatment_time = 1991)

 summary(fitMultiSynth.out)
 }
}
\author{
Dalston G. Ward: \email{ward.dalston@wustl.edu}
}
\references{
\itemize{
\item Abadie, A., Diamond, A., Hainmueller, J. (2010). Synthetic Control Methods for Comparative Case Studies: Estimating the Effect of California's Tobacco Control Program. \emph{Journal of the American Statistical Association} 105 (490) 493-505.
\item Abadie, A., Diamond, A., Hainmueller, J. (2015). Comparative Politics and the Synthetic Control Method.  \emph{American Journal of Political Science} 59 (2) 495-510
}
}
\seealso{
\code{\link{MultiSynth}}

\code{\link{fitMultiSynth}}
}

