% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MultiSynthGapsPlot.R
\docType{methods}
\name{gaps.plot,MultiSynth-method}
\alias{gaps.plot,MultiSynth-method}
\title{Gaps plots for MultiSynth Objects}
\usage{
gaps.plot(synth.res = NA,
 dataprep.res = NA,
 Ylab = c("Title"),
 Xlab = c("Time"),
 Main = c("Gaps: Treated - Synthetic"),
 tr.intake = NA,
 Ylim = NA, Z.plot = FALSE)
}
\arguments{
\item{synth.res}{An object of class \code{\link{MultiSynth}}}

\item{dataprep.res}{NA.  Not used for MultiSynth gaps plotting, should be left as NA.}

\item{Ylab}{Optional label for Y axis}

\item{Xlab}{Optional label for X axis}

\item{Main}{Optional main title}

\item{tr.intake}{NA. Not used for MultiSynth gaps plotting, should be left as NA.}

\item{Ylim}{Optional Y axis limits}

\item{Z.plot}{Logical. If TRUE only the pre-treatment period is plotted.}
}
\description{
Plots the gaps in the trajectory of the outcome variable for the treated case and its synthetic control for each fit in a MultiSynth object.  The main analysis is plotted as a solid black line, and each of the placebo or leave-one-out fits is plotted as a light gray solid line.
}
\details{
The user need only prvide an object of class \code{MultiSynth} as the arguement to synth.res for the function to work. Two other arguments, dataprep.res and tr.intake should be left as NA.  Other arguments are optional.
}
\examples{
##Example: Hainmueller and Diamond's Toy panel dataset

 #load data
data(synth.data)

## create matrices from panel data that provide inputs for fitMultiSynth()
dataprep.out<-
 dataprep(
   foo = synth.data,
   predictors = c("X1", "X2", "X3"),
   predictors.op = "mean",
   dependent = "Y",
   unit.variable = "unit.num",
   time.variable = "year",
   special.predictors = list(
     list("Y", 1991, "mean"),
     list("Y", 1985, "mean"),
     list("Y", 1980, "mean")
   ),
   treatment.identifier = 7,
   controls.identifier = c(29, 2, 13, 17, 32, 38),
   time.predictors.prior = c(1984:1989),
   time.optimize.ssr = c(1984:1990),
   unit.names.variable = "name",
   time.plot = 1984:1996
 )

 ## Fit a multisynth object
 fitMultiSynth.out <- fitMultiSynth(dataprep.out, treatment_time = 1991)

 gaps.plot(fitMultiSynth.out)
}
\author{
Dalston G. Ward: \email{ward.dalston@wustl.edu}
}
\references{
\itemize{
\item Abadie, A., Diamond, A., Hainmueller, J. (2010). Synthetic Control Methods for Comparative Case Studies: Estimating the Effect of California's Tobacco Control Program. \emph{Journal of the American Statistical Association} 105 (490) 493-505.
\item Abadie, A., Diamond, A., Hainmueller, J. (2015). Comparative Politics and the Synthetic Control Method.  \emph{American Journal of Political Science} 59 (2) 495-510
}
}
\seealso{
\code{\link{fitMultiSynth}}

\code{\link{MultiSynth}}

\code{\link{plot,MultiSynth-method}}

\code{\link{summary,MultiSynth-method}}

\code{\link{path.plot,MultiSynth-method}}

\code{\link{gaps.plot}}
}

