% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getFunctions.R
\docType{methods}
\name{getPreps}
\alias{getPreps}
\alias{getPreps,MultiSynth-method}
\title{Extract the Data Matrices for All Cases in a MultiSynth Analysis}
\usage{
getPreps(input)

\S4method{getPreps}{MultiSynth}(input)
}
\arguments{
\item{input}{An object of class "MultiSynth"}
}
\value{
A list of lists. Each list contains the six dataprep matrices for a single case.
}
\description{
Extracts from a MultiSynth object the named list containing the X0, X1, Z0, Z1, Y0plot, and Y1plot matrices for each case. Each element of this list is a list containing the six relevent matrices for each placebo/leave-one-out case.
}
\author{
Dalston G. Ward: \email{ward.dalston@wustl.edu}
}
\seealso{
\code{\link{MultiSynth}}

\code{\link{getInput}}

\code{\link{getFits}}

\code{\link{getTreated}}

\code{\link{getTreatmentTime}}

\code{\link{getCase}}

\code{\link{getStats}}
}

