% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getFunctions.R
\docType{methods}
\name{getStats}
\alias{getStats}
\alias{getStats,MultiSynth-method}
\title{Extract the statistics for estimated synthetic control fits in a MultiSynth object}
\usage{
getStats(input)

\S4method{getStats}{MultiSynth}(input)
}
\arguments{
\item{input}{An object of class "MultiSynth"}
}
\value{
A matrix containing the statistics for the MultiSynth cases and main case.
}
\description{
Returns a matrix containing the pretreatment RMSPE, posttreatment RMSPE, post- to pre-treatment RMSPE ratio, covariate loss, and ATE for each of the placebo/leave-one-out cases and the main case in a MultiSynth analysis.
}
\author{
Dalston G. Ward: \email{ward.dalston@wustl.edu}
}
\seealso{
\code{\link{MultiSynth}}

\code{\link{getInput}}

\code{\link{getPreps}}

\code{\link{getFits}}

\code{\link{getTreated}}

\code{\link{getTreatmentTime}}

\code{\link{getCase}}
}

