% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getFunctions.R
\name{getTreated}
\alias{getTreated}
\title{Extract information about which unit is the treatment unit in a MultiSynth analysis}
\usage{
getTreated(input)
}
\arguments{
\item{input}{An object of class "MultiSynth"}
}
\value{
A character vector containing the name and number of hte treated unit
}
\description{
Returns a vector containing the name and unit number of the treated unit in a MultiSynth object.
}
\author{
Dalston G. Ward: \email{ward.dalston@wustl.edu}
}
\seealso{
\code{\link{MultiSynth}}

\code{\link{getInput}}

\code{\link{getPreps}}

\code{\link{getFits}}

\code{\link{getTreatmentTime}}

\code{\link{getCase}}

\code{\link{getStats}}
}

