% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getFunctions.R
\docType{methods}
\name{getTreatmentTime}
\alias{getTreatmentTime}
\alias{getTreatmentTime,MultiSynth-method}
\title{Extract the Time Treatment is Administered in a MultiSynth Analysis}
\usage{
getTreatmentTime(input)

\S4method{getTreatmentTime}{MultiSynth}(input)
}
\arguments{
\item{input}{An object of class "MultiSynth"}
}
\value{
A scalar containing the time treatment is administered
}
\description{
Returns the time treatment is administered in a MultiSynth analysis.
}
\author{
Dalston G. Ward: \email{ward.dalston@wustl.edu}
}
\seealso{
\code{\link{MultiSynth}}

\code{\link{getInput}}

\code{\link{getPreps}}

\code{\link{getFits}}

\code{\link{getTreated}}

\code{\link{getCase}}

\code{\link{getStats}}
}

